Configures database and improves Docker setup
Sets up a MySQL database service within the Docker environment for local development.

Adds healthcheck to the database service to ensure it's running before the application starts.

Configures the application to depend on the database service being healthy before starting.

Moves the database service definition to the top of the docker-compose file for readability.

Includes essential composer commands in the Dockerfile to optimize the application's autoloading and performance.

Also generates the app key, and updates file permissions for storage and cache directories.